OTel.SDK.Trace.Internal

Disclaimer

In general, changes to this module will not be reflected in the library's version updates. Direct use of this module should be done with utmost care, otherwise invariants will easily be violated.

data TracerProviderSpec

Constructors

TracerProviderSpec

Fields

tracerProviderSpecNow :: IO Timestamp

tracerProviderSpecLogger :: Logger

tracerProviderSpecSeed :: Seed

tracerProviderSpecIdGenerator :: forall a. Logger -> (IdGeneratorSpec -> IO a) -> IO a

tracerProviderSpecSpanProcessors :: forall a. [Logger -> (SpanProcessorSpec -> IO a) -> IO a]

tracerProviderSpecSampler :: forall a. Logger -> (SamplerSpec -> IO a) -> IO a

tracerProviderSpecResource :: Resource Attrs

tracerProviderSpecSpanAttrsLimits :: AttrsLimits 'AttrsForSpan

tracerProviderSpecSpanEventAttrsLimits :: AttrsLimits 'AttrsForSpanEvent

tracerProviderSpecSpanLinkAttrsLimits :: AttrsLimits 'AttrsForSpanLink

tracerProviderSpecCallStackAttrs :: CallStack -> AttrsBuilder 'AttrsForSpan

tracerProviderSpecSpanContextMeta :: SpanContext -> [Pair]

defaultTracerProviderSpec :: TracerProviderSpec

withTracerProvider :: forall m a. MonadUnliftIO m => TracerProviderSpec -> (TracerProvider -> m a) -> m a

withTracerProviderIO :: forall a. TracerProviderSpec -> (TracerProvider -> IO a) -> IO a

data SpanProcessor

Constructors

SpanProcessor

Fields

spanProcessorOnSpanStart :: Context -> (UpdateSpanSpec -> IO (Span Attrs)) -> IO ()

spanProcessorOnSpanEnd :: Span Attrs -> IO ()

spanProcessorShutdown :: IO ()

spanProcessorForceFlush :: IO ()

buildSpanProcessor :: forall a. Resource Attrs -> Logger -> SpanProcessorSpec -> (SpanProcessor -> IO a) -> IO a

data SimpleSpanProcessorSpec

Constructors

SimpleSpanProcessorSpec

Fields

simpleSpanProcessorSpecName :: Text

simpleSpanProcessorSpecExporter :: forall a. Logger -> (SpanExporterSpec -> IO a) -> IO a

simpleSpanProcessorSpecOnSpansExported :: OnSpansExported ()

defaultSimpleSpanProcessorSpec :: SimpleSpanProcessorSpec

simpleSpanProcessor :: forall a. SimpleSpanProcessorSpec -> Logger -> (SpanProcessorSpec -> IO a) -> IO a

data SpanProcessorSpec

Constructors

SpanProcessorSpec

Fields

spanProcessorSpecName :: Text

spanProcessorSpecExporter :: forall a. Logger -> (SpanExporterSpec -> IO a) -> IO a

spanProcessorSpecOnSpanStart :: Context -> (UpdateSpanSpec -> SpanProcessorM (Span Attrs)) -> SpanProcessorM ()

spanProcessorSpecOnSpanEnd :: Span Attrs -> SpanProcessorM ()

spanProcessorSpecShutdown :: SpanProcessorM ()

spanProcessorSpecShutdownTimeout :: Int

spanProcessorSpecForceFlush :: SpanProcessorM ()

spanProcessorSpecForceFlushTimeout :: Int

spanProcessorSpecOnTimeout :: OnTimeout ()

spanProcessorSpecOnException :: OnException ()

defaultSpanProcessorSpec :: SpanProcessorSpec

data SpanExportResult

Constructors

SpanExportResultSuccess

SpanExportResultFailure

data SpanExporter

Constructors

SpanExporter

Fields

spanExporterExport :: Batch (Span Attrs) -> (SpanExportResult -> IO ()) -> IO ()

spanExporterShutdown :: IO ()

spanExporterForceFlush :: IO ()

buildSpanExporter :: Resource Attrs -> Logger -> SpanExporterSpec -> IO SpanExporter

data OTLPSpanExporterSpec

Constructors

OTLPSpanExporterSpec

Fields

otlpSpanExporterSpecManager :: Manager

otlpSpanExporterSpecEndpoint :: URI

otlpSpanExporterSpecTimeout :: Int

otlpSpanExporterSpecProtocol :: OTLPProtocol

otlpSpanExporterSpecHeaders :: [Header]

A list of headers to include when communicating with the observability backend (e.g. Honeycomb) over HTTP.

Use this list to include the necessary secrets for talking with your observability backend(s).

otlpSpanExporterSpecRedactedRequestHeaders :: [HeaderName]

A list of sensitive header names that will be redacted before a Request is displayed. Note that the only time a Request is displayed is when the span exporter encounters an HttpException when communicating with the observability backend (e.g. Honeycomb). The default is mempty.

Use this list to avoid leaking sensitive data like API keys into your logs:

defaultOTLPSpanExporterSpec
  { otlpSpanExporterSpecRedactedRequestHeaders = ["x-honeycomb-team"]
  }

otlpSpanExporterSpecRedactedResponseHeaders :: [HeaderName]

A list of sensitive header names that will be redacted before a Response is displayed. Note that the only time a Response is displayed is when the span exporter encounters an HttpException when communicating with the observability backend (e.g. Honeycomb). The default is mempty.

Use this list to avoid leaking sensitive data like API keys into your logs.

defaultOTLPSpanExporterSpec
  { otlpSpanExporterSpecRedactedResponseHeaders = ["x-honeycomb-team"]
  }

otlpSpanExporterSpecWorkerQueueSize :: Int

otlpSpanExporterSpecWorkerCount :: Int

otlpSpanExporterSpecRetryPolicy :: RetryPolicyM IO

The retry policy to use when communicating with the observability backend produces an exception.

The default is defined as follows:

fullJitterBackoff 10_000 <> limitRetries 10

simulatePolicyPP can be used to get an idea of the retry policy's iterations and total cumulative delay:

ghci> simulatePolicyPP 10 $ fullJitterBackoff 10000 <> limitRetries 10
0: 6.659ms
1: 12.302ms
2: 21.228ms
3: 45.048ms
4: 128.142ms
5: 274.269ms
6: 351.933ms
7: 688.239ms
8: 1313.14ms
9: 4806.224ms
10: Inhibit
Total cumulative delay would be: 7647.184ms

For more info on defining custom retry policies, see Control.Retry.

defaultOTLPSpanExporterSpec :: OTLPSpanExporterSpec

otlpSpanExporter :: forall a. OTLPSpanExporterSpec -> Logger -> (SpanExporterSpec -> IO a) -> IO a

data OTLPProtocol

Constructors

OTLPProtocolHTTPProtobuf

httpProtobufProtocol :: OTLPProtocol

data OTLPSpanExporterItem

An ad-hoc helper type for use in otlpSpanExporterIO.

Constructors

OTLPSpanExporterItem

Fields

otlpSpanExporterItemBatch :: Batch (Span Attrs)

otlpSpanExporterItemCallback :: SpanExportResult -> IO ()

otlpSpanExporterResource :: Resource Attrs

stmSpanExporter :: forall a. TMQueue (Span Attrs) -> Logger -> (SpanExporterSpec -> IO a) -> IO a

data SpanExporterSpec

Constructors

SpanExporterSpec

Fields

spanExporterSpecName :: Text

spanExporterSpecExport :: Batch (Span Attrs) -> (SpanExportResult -> IO ()) -> SpanExporterM ()

spanExporterSpecShutdown :: SpanExporterM ()

spanExporterSpecShutdownTimeout :: Int

spanExporterSpecForceFlush :: SpanExporterM ()

spanExporterSpecForceFlushTimeout :: Int

spanExporterSpecOnTimeout :: OnTimeout ()

spanExporterSpecOnException :: OnException ()

defaultSpanExporterSpec :: SpanExporterSpec

data Sampler

Constructors

Sampler

Fields

samplerName :: Text

samplerDescription :: Text

samplerShouldSample :: SamplerInput -> IO SamplingResult

buildSampler :: forall m. MonadIO m => Logger -> SamplerSpec -> m Sampler

data SamplerSpec

Constructors

SamplerSpec

Fields

samplerSpecName :: Text

samplerSpecDescription :: Text

samplerSpecShouldSample :: SamplerInput -> SamplerM SamplingResult

samplerSpecOnException :: OnException SamplingResult

defaultSamplerSpec :: SamplerSpec

alwaysOnSampler :: forall a. Logger -> (SamplerSpec -> IO a) -> IO a

alwaysOffSampler :: forall a. Logger -> (SamplerSpec -> IO a) -> IO a

data ParentBasedSamplerSpec

Constructors

ParentBasedSamplerSpec

Fields

parentBasedSamplerSpecOnRoot :: forall a. Logger -> (SamplerSpec -> IO a) -> IO a

parentBasedSamplerSpecOnRemoteParentSampled :: forall a. Logger -> (SamplerSpec -> IO a) -> IO a

parentBasedSamplerSpecOnRemoteParentNotSampled :: forall a. Logger -> (SamplerSpec -> IO a) -> IO a

parentBasedSamplerSpecOnLocalParentSampled :: forall a. Logger -> (SamplerSpec -> IO a) -> IO a

parentBasedSamplerSpecOnLocalParentNotSampled :: forall a. Logger -> (SamplerSpec -> IO a) -> IO a

defaultParentBasedSamplerSpec :: ParentBasedSamplerSpec

parentBasedSampler :: forall a. ParentBasedSamplerSpec -> Logger -> (SamplerSpec -> IO a) -> IO a

constDecisionSampler :: SamplingDecision -> SamplerSpec

data SamplerInput

Constructors

SamplerInput

Fields

samplerInputContext :: Context

samplerInputTraceId :: TraceId

samplerInputSpanName :: SpanName

samplerInputSpanKind :: SpanKind

samplerInputSpanAttrs :: AttrsBuilder 'AttrsForSpan

samplerInputSpanLinks :: SpanLinks AttrsBuilder

data SamplingResult

Constructors

SamplingResult

Fields

samplingResultDecision :: SamplingDecision

samplingResultSpanAttrs :: AttrsBuilder 'AttrsForSpan

samplingResultTraceState :: TraceState

defaultSamplingResult :: SamplingResult

data SamplingDecision

Constructors

SamplingDecisionDrop

SamplingDecisionRecordOnly

SamplingDecisionRecordAndSample

samplingDecisionDrop :: SamplingDecision

samplingDecisionRecordOnly :: SamplingDecision

samplingDecisionRecordAndSample :: SamplingDecision

newtype SpanProcessorM a

Constructors

SpanProcessorM

Fields

unSpanProcessorM :: SpanExporter -> LoggingT IO a

askSpanExporter :: SpanProcessorM SpanExporter

runSpanProcessorM :: SpanExporter -> Logger -> OnTimeout a -> OnException a -> Int -> [SeriesElem] -> SpanProcessorM a -> IO a

newtype SpanExporterM a

Constructors

SpanExporterM

Fields

unSpanExporterM :: Resource Attrs -> LoggingT IO a

askResource :: SpanExporterM (Resource Attrs)

runSpanExporterM :: Resource Attrs -> Logger -> OnTimeout a -> OnException a -> Int -> [SeriesElem] -> SpanExporterM a -> IO a

newtype SamplerM a

Constructors

SamplerM

Fields

unSamplerM :: LoggingT IO a

runSamplerM :: Logger -> OnException a -> [SeriesElem] -> SamplerM a -> IO a

newtype IdGeneratorM a

Constructors

IdGeneratorM

Fields

unIdGeneratorM :: PRNG -> LoggingT IO a

runIdGeneratorM :: PRNG -> Logger -> IdGeneratorM a -> IO a

data IdGenerator

Constructors

IdGenerator

Fields

idGeneratorGenTraceId :: PRNG -> IO TraceId

idGeneratorGenSpanId :: PRNG -> IO SpanId

buildIdGenerator :: forall m. MonadIO m => Logger -> IdGeneratorSpec -> m IdGenerator

data IdGeneratorSpec

Constructors

IdGeneratorSpec

Fields

idGeneratorSpecName :: Text

idGeneratorSpecGenTraceId :: IdGeneratorM TraceId

idGeneratorSpecGenSpanId :: IdGeneratorM SpanId

defaultIdGeneratorSpec :: IdGeneratorSpec

newtype PRNG

Constructors

PRNG

Fields

unPRNG :: GenIO

genUniform :: forall a. Variate a => IdGeneratorM a

newPRNGRef :: Seed -> IO (MVar PRNG)

newtype OnException a

Constructors

OnException

Fields

runOnException :: SomeException -> [SeriesElem] -> LoggingT IO a

askException :: OnException SomeException

askExceptionMetadata :: OnException [SeriesElem]

newtype OnTimeout a

Constructors

OnTimeout

Fields

runOnTimeout :: Int -> [SeriesElem] -> LoggingT IO a

askTimeoutMicros :: OnTimeout Int

askTimeoutMetadata :: OnTimeout [SeriesElem]

newtype OnSpansExported a

Constructors

OnSpansExported

Fields

runOnSpansExported :: Batch (Span Attrs) -> SpanExportResult -> [SeriesElem] -> LoggingT IO a

askSpansExported :: OnSpansExported (Batch (Span Attrs))

askSpansExportedResult :: OnSpansExported SpanExportResult

askSpansExportedMetadata :: OnSpansExported [SeriesElem]

newtype Batch a

Constructors

Batch

Fields

unBatch :: [a]

singletonBatch :: a -> Batch a

fromListBatch :: [a] -> Batch a

data ConcurrentWorkersSpec item

Constructors

ConcurrentWorkersSpec

Fields

concurrentWorkersSpecQueueSize :: Int

concurrentWorkersSpecWorkerCount :: Int

concurrentWorkersSpecProcessItem :: item -> IO ()

concurrentWorkersSpecLogger :: Logger

concurrentWorkersSpecLoggingMeta :: [SeriesElem]

concurrentWorkersSpecOnException :: item -> OnException ()

defaultConcurrentWorkersSpec :: ConcurrentWorkersSpec item

data ConcurrentWorkers item

Constructors

ConcurrentWorkers

Fields

concurrentWorkersEnqueueItem :: item -> IO ()

concurrentWorkersStopWorkers :: IO ()

withConcurrentWorkers :: forall item a. (ToJSON item, Typeable item) => ConcurrentWorkersSpec item -> (ConcurrentWorkers item -> IO a) -> IO a

unlessSTM :: Monoid a => STM Bool -> STM (IO a) -> IO a

withAll :: forall a b. [(a -> b) -> b] -> ([a] -> b) -> b

defaultSystemSeed :: Seed

defaultManager :: Manager

spanSummary :: Span attrs -> Value

redactHttpExceptionHeaders :: Set HeaderName -> Set HeaderName -> SomeException -> SomeException