OTel.SDK.Trace.SpanExporter.OTLP

Documentation

otlpSpanExporter :: forall a. OTLPSpanExporterSpec -> Logger -> (SpanExporterSpec -> IO a) -> IO a

data OTLPSpanExporterSpec

defaultOTLPSpanExporterSpec :: OTLPSpanExporterSpec

otlpSpanExporterSpecManager :: OTLPSpanExporterSpec -> Manager

otlpSpanExporterSpecEndpoint :: OTLPSpanExporterSpec -> URI

otlpSpanExporterSpecTimeout :: OTLPSpanExporterSpec -> Int

otlpSpanExporterSpecProtocol :: OTLPSpanExporterSpec -> OTLPProtocol

otlpSpanExporterSpecHeaders :: OTLPSpanExporterSpec -> [Header]

A list of headers to include when communicating with the observability backend (e.g. Honeycomb) over HTTP.

Use this list to include the necessary secrets for talking with your observability backend(s).

otlpSpanExporterSpecRedactedRequestHeaders :: OTLPSpanExporterSpec -> [HeaderName]

A list of sensitive header names that will be redacted before a Request is displayed. Note that the only time a Request is displayed is when the span exporter encounters an HttpException when communicating with the observability backend (e.g. Honeycomb). The default is mempty.

Use this list to avoid leaking sensitive data like API keys into your logs:

defaultOTLPSpanExporterSpec
  { otlpSpanExporterSpecRedactedRequestHeaders = ["x-honeycomb-team"]
  }

otlpSpanExporterSpecRedactedResponseHeaders :: OTLPSpanExporterSpec -> [HeaderName]

A list of sensitive header names that will be redacted before a Response is displayed. Note that the only time a Response is displayed is when the span exporter encounters an HttpException when communicating with the observability backend (e.g. Honeycomb). The default is mempty.

Use this list to avoid leaking sensitive data like API keys into your logs.

defaultOTLPSpanExporterSpec
  { otlpSpanExporterSpecRedactedResponseHeaders = ["x-honeycomb-team"]
  }

otlpSpanExporterSpecWorkerQueueSize :: OTLPSpanExporterSpec -> Int

otlpSpanExporterSpecWorkerCount :: OTLPSpanExporterSpec -> Int

otlpSpanExporterSpecRetryPolicy :: OTLPSpanExporterSpec -> RetryPolicyM IO

The retry policy to use when communicating with the observability backend produces an exception.

The default is defined as follows:

fullJitterBackoff 10_000 <> limitRetries 10

simulatePolicyPP can be used to get an idea of the retry policy's iterations and total cumulative delay:

ghci> simulatePolicyPP 10 $ fullJitterBackoff 10000 <> limitRetries 10
0: 6.659ms
1: 12.302ms
2: 21.228ms
3: 45.048ms
4: 128.142ms
5: 274.269ms
6: 351.933ms
7: 688.239ms
8: 1313.14ms
9: 4806.224ms
10: Inhibit
Total cumulative delay would be: 7647.184ms

For more info on defining custom retry policies, see Control.Retry.

data OTLPProtocol

httpProtobufProtocol :: OTLPProtocol